+++ cmake/CompileOptionSelector.cmake
@@ -115,8 +115,14 @@ endmacro()
 
 
 macro(precision)
-  if(real_type STREQUAL "float")
-  else()
+  if(real_type STREQUAL "OFF")
+  # nothing, default is float
+  set(real_type "float")
+
+  elseif(real_type STREQUAL "float")
+  # nothing
+
+  elseif(real_type STREQUAL "double")
     ADD_DEFINITIONS(-D_REAL_IS_DOUBLE_)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -D_REAL_IS_DOUBLE_")
 
@@ -133,5 +139,9 @@ macro(precision)
 
     endif()
 
+  else() # neither 'float' nor 'double'
+    message("@@@@@@@@@@@")
+    message("FATAL ERROR : Invalid floating type : ${real_type}")
+    message("@@@@@@@@@@@")
   ENDIF()
 endmacro()